Interpreter: Add "in" and "not in" operators
Closes: #3600
